@@ -25,14 +25,13 @@ ext {
 }
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion "27.0.3"
+    compileSdkVersion 28
 
     defaultConfig {
         minSdkVersion 17
-        targetSdkVersion 27
-        versionCode 4
-        versionName "1.3.1"
+        targetSdkVersion 28
+        versionCode 5
+        versionName "1.4.0"
     }
     buildTypes {
         release {
@@ -43,10 +42,10 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support:support-annotations:27.1.1'
-    implementation 'com.android.support:support-v4:27.1.1'
-    implementation 'com.github.dcendents:android-maven-gradle-plugin:2.0'
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'androidx.annotation:annotation:1.1.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'com.github.dcendents:android-maven-gradle-plugin:2.1'
 }
 // build a jar with source files
 task sourcesJar(type: Jar) {
