@@ -3,7 +3,8 @@ package com.mohamadamin.persianmaterialdatetimepicker;
 
 import android.content.Context;
 import android.graphics.Typeface;
-import android.support.v4.util.SimpleArrayMap;
+
+import androidx.collection.SimpleArrayMap;
 
 public class TypefaceHelper {
 
